Add files via upload
This is a project I did that created a Graphical User Interface for a Course Scheduler. It uses a database of class information. Allow user to select specific semester, add course, and then the student is added to the database using Student ID.  It then schedules the courses and displays the users course schedule. 

More detailed Explanation:

Add Semester:
A semester is added to the database. The semester is identified by one name.
Add Course:
A new course is added to the database. The course is identified by the semester it is
to be added to, the code for the course, the description of the course and the maximum
number of students the course will contain that semester.
Add Student:
A student is added to the database. The student is identified by a studentID, the
student’s first name, and the student’s last name.
This phase of the project will implement the following Student functions:
Schedule Course:
The student will be scheduled in the class for the specified semester, if there are
seats available. If there are no seats available, the student will be put a wait list for that
Course. The waiting list must be maintained in the order the students tried to schedule
the course.
Display Schedule:
The Display Schedule function will display the current schedule for a specified
student for the current semester.
Display Courses:
The Display Courses function will display a complete list of courses for the current
semester.
Testing scenario:
A testing scenario will be provided to assist you in testing this application. It will be
called Final Project Part 1 Test Script in Canvas.
Database considerations:
Your database will be created but all of the tables should be empty when your
project is submitted.