Add files via upload
This is my final project for Data Management for Data Science. I used MongoDB and SQL to understand data given by FIFA. I answered my questions by using these two programs and wrote about how the data can be used to help teams in the future. And also how the analysis of data is different in MongoDB compared to SQL.